PUR-6212: Simplify generator API and remove broken validator
This is part 1 of a series of pull requests. Originally written by Igor Sechyn.

The JWT validator in this library doesn't follow the ASAP spec, and is
missing crucial functionality. In this PR I remove it. The JWT generator
API is also changed slightly so that it is not tied to the validator.

The next PR will re-add validator code which follows the ASAP spec and
supports useful features like mirrored HTTP public key repos with
caching.